combine request into one namespace

+++ GoLoginTools.POCO/CanvasModel.cs
@@ -5,6 +5,6 @@ namespace GoLoginTools.POCO
 	public class CanvasModel
 	{
         public string mode { get; set; } = CanvasMode.off.ToString();
-        public int noise { get; set; }
+        public double noise { get; set; }
     }
 }
