+++ GoLoginTools.POCO/WebGlModel.cs
@@ -6,6 +6,6 @@ namespace GoLoginTools.POCO
 	{
 		public string mode { get; set; } = WebGlMode.noise.ToString();
 		public int getClientRectNoise { get; set; } = 0;
-		public int noise { get; set; } = 0;
+		public double noise { get; set; } = 0.0d;
     }
 }
